Mudança nos models de BigDecimal para int

+++ src/main/java/br/com/promocoes_api/api/dto/produto/ProdutoDetailsDTO.java
@@ -1,13 +1,12 @@
 package br.com.promocoes_api.api.dto.produto;
 
-import java.math.BigDecimal;
 import java.time.LocalDate;
 
 public record ProdutoDetailsDTO(
     Long id, 
     String nome, 
     String descricao, 
-    BigDecimal preco, 
+    int preco, 
     LocalDate dataValidade, 
     LocalDate dataUltimaAtualizacaoPreco, 
     String marca,
