+++ src/main/java/br/com/promocoes_api/api/dto/promocao/PromocaoRequestDTO.java
@@ -1,11 +1,11 @@
 package br.com.promocoes_api.api.dto.promocao;
 
-import java.math.BigDecimal;
+
 import java.time.LocalDate;
 
 public record PromocaoRequestDTO(
     String titulo,
-    BigDecimal desconto,
+    int desconto,
     LocalDate dataInicioPromo,
     LocalDate dataFimPromo
 ) {}
